Update EXPLAINER.md: Some small changes that could improve readability (#26)
* Update EXPLAINER.md: Use 'let' instead of 'var' keyword

'const' is used in examples so ES6 is assumed.  Prefer 'let'
to 'var' since 'let' is lexically scoped to the closest closing
block expression.

* Update EXPLAINER.md: Pull apart some calls for readability, add some args

This commit is intended to make clear what calls are awaited,
when some calls will return, and where data is expected to be
after a handler is dispatched.

* Update EXPLAINER.md: Move write block into onload handler

The text flow suggests the read result is mutated and the
changed result is written back to the file.  Move the block
which writes back to within the read handler.