Add options to createWriter() to deal with atomic writes. (#51)
* Add options to createWriter() to deal with atomic writes.

This attempts to solve #37 by making it explicit that changes made to files
might not be reflected on disk until close() is called on the
writer.

This also introduces a "atomic" mode for FileSystemWriter where any changes
are written to a temporary file, and only on close is that file
moved on top of the destination file. This should address #48.