Remove or annotate features that aren't implemented yet.
This makes it hopefully clearer what the current shape of the API is
like.

In particular this removes the "inPlace" mode of writers. We probably
want something like that, but it is not clear yet what such an API will
look like, so it is less confusing if the spec only describes atomic
writes.