chore(rebuild): Meta: Add cloud file handling proposal (#411)
This proposal adds a a new `getCloudIdentifiers()` method to `FileSystemHandle`, which allows to retrieve cloud handles for a file/directory. A cloud handle consists of a vendor identifiert (e.g. "drive.google.com") and a file identifier. With these, the web app can talk to the cloud storage provider through its APIs directly to retrieve/modify the file. This is useful for web apps to figure out if a file is already backed by cloud storage and allows for easier transfer across machines as just the identifier instead of the entire file contents need to be transfered.

SHA: 4d4fbd84aaee7cc54696cd9128960415601ddd60
Reason: push, by @a-sully

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>